worker-service/pyproject.toml
[project]
name = "worker-service"
version = "0.1.0"
description = "Phase 1 worker service for asynchronous task execution"
requires-python = ">=3.13"
dependencies = [
  "aio-pika>=9.5.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "greenlet>=3.2.0",
  "pillow>=11.2.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.8.0",
  "pypdf>=5.4.0",
  "resend>=2.13.0",
  "sqlalchemy>=2.0.40",
  "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
